refactor: add purchase-selection as a module
