+++ src/stacks-hierarchy/Root_PurchaseConfirmationScreen/navigation-types.ts
@@ -1,5 +1,5 @@
 import {TicketRecipientType} from '@atb/ticketing';
-import {PurchaseSelectionType} from '@atb/purchase-selection';
+import {PurchaseSelectionType} from '@atb/modules/purchase-selection';
 
 export type Root_PurchaseConfirmationScreenParams = {
   selection: PurchaseSelectionType;
