+++ src/stacks-hierarchy/Root_PurchaseOverviewScreen/components/ProductSelectionByAlias.tsx
@@ -19,7 +19,7 @@ import {onlyUniquesBasedOnField} from '@atb/utils/only-uniques';
 import {
   type PurchaseSelectionType,
   usePurchaseSelectionBuilder,
-} from '@atb/purchase-selection';
+} from '@atb/modules/purchase-selection';
 
 type Props = {
   color: InteractiveColor;
