+++ src/stacks-hierarchy/Root_PurchaseOverviewScreen/components/Travellers/use-user-count-state.ts
@@ -4,7 +4,7 @@ import {UserCountState} from './types';
 import {
   useSelectableUserProfiles,
   type PurchaseSelectionType,
-} from '@atb/purchase-selection';
+} from '@atb/modules/purchase-selection';
 
 type ReducerState = {
   userProfilesWithCount: UserProfileWithCount[];
