+++ src/stacks-hierarchy/Root_PurchaseOverviewScreen/navigation-types.ts
@@ -1,4 +1,4 @@
-import type {PurchaseSelectionType} from '@atb/purchase-selection';
+import type {PurchaseSelectionType} from '@atb/modules/purchase-selection';
 
 export type Root_PurchaseOverviewScreenParams = {
   selection: PurchaseSelectionType;
